pyproject.toml
[project]
name = "nexus-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "flask>=3.1.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.8",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "pytube>=15.0.0",
    "sentence-transformers>=5.2.2",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.2.4",
]
